moved influx to handler
